+++ CHANGELOG.md
@@ -5,6 +5,9 @@ IMPROVEMENTS:
  * plugin/target/aws: Prevent scaling if an instance refresh is in progress [[GH-597](https://github.com/hashicorp/nomad-autoscaler/pull/597)]
  * plugin/target/aws: Add new configuration `retry_attempts` to account for potentially slow ASG update operations [[GH-594](https://github.com/hashicorp/nomad-autoscaler/pull/594)
 
+BUG FIXES:
+ * plugin/apm/datadog: Fixed a panic when Datadog queries return `null` values [[GH-606](https://github.com/hashicorp/nomad-autoscaler/pull/606)]
+
 ## 0.3.7 (June 10, 2022)
 
 IMPROVEMENTS: